// src/global/types/classTest.types.ts
export interface ClassTest {
    autoId: number;
    scheduleId: string;
    testTitle: string;
    questionType: string;
    className: string;
    testDateTime: string;
    subject: string;
    schoolName: string;
    id: number;
    testId: number;
    totalQuestion: number;
    totalTime: number;
    schoolId: number;
    testType: number;
    totalMark: number;
    markCorrectAns: number;
    markIncorrectAns: number;
    markNotAttempt: number;
}